Module LZ4

LZ4 compression algorithm

exception Input_too_large

exception Corrupted

val compress_bound : int -> int

compress_bound sz returns the maximum size of the LZ4 output for sz bytes of input. It raises Input_too_large if sz > 0x7E000000 or if the result would be larger than Pervasives.max_int (the latter is only possible, if Sys.word_size = 32). Raises Invalid_argument "LZ4.compress_bound" if sz is negative.

val version_number : int

Version number of the LZ4 library

module type S = sig ... end

module Bytes : S with type storage = Stdlib.Bytes.t

module Bigbytes : S with type storage = (char, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t